Extend packet_id to 64 bits in SentPacket and PacketOptions to support QUIC packet numbers.
Bug: webrtc:9103

+++ rtc_base/asyncpacketsocket.h
@@ -42,7 +42,7 @@ struct PacketOptions {
   ~PacketOptions();
 
   DiffServCodePoint dscp = DSCP_NO_CHANGE;
-  int packet_id = -1;  // 16 bits, -1 represents "not set".
+  int64_t packet_id = -1;  // 16 bits, -1 represents "not set".
   PacketTimeUpdateParams packet_time_params;
   // PacketInfo is passed to SentPacket when signaling this packet is sent.
   PacketInfo info_signaled_after_sent;
